+++ src/features/project/post/pages/CreatePostDrawer.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import {
   Drawer,
   Box,
@@ -8,43 +8,31 @@ import {
   Paper,
   Divider,
   TextField,
-  Button,
   CircularProgress,
   FormControl,
   InputLabel,
   Select,
   MenuItem,
   Tooltip,
-  Chip,
-  LinearProgress,
-  Alert,
-  Modal,
-  Avatar,
 } from "@mui/material";
 import CloseIcon from "@mui/icons-material/Close";
-import {
-  InfoOutlined,
-  CloudUpload,
-  Delete,
-  ZoomIn,
-  Refresh,
-} from "@mui/icons-material";
+import { InfoOutlined, CloudUpload } from "@mui/icons-material";
 import { useTheme } from "@mui/material/styles";
 import { useDispatch, useSelector } from "react-redux";
 import { useParams } from "react-router-dom";
 import {
-  createPostId,
   createPost,
-  deleteAttachment,
   bulkActivateAttachments,
-  cleanupPostAttachments,
 } from "@/features/project/post/postSlice";
-import * as postAPI from "@/api/post";
 import CustomButton from "@/components/common/customButton/CustomButton";
-import FilePreviewModal from "./FilePreviewModal";
-import FileAttachmentCard from "./FileAttachmentCard";
-import { validateFile } from "@/utils/validateFile";
-import { uploadFileWithPresignedUrl } from "@/utils/uploadFileWithPresignedUrl";
+import FilePreviewModal from "../components/FilePreviewModal";
+import FileAttachmentCard from "../components/FileAttachmentCard";
+import usePostForm from "@/hooks/usePostForm";
+import {
+  hasUnfinishedUploads,
+  hasFailedUploads,
+  getSuccessfulPostAttachmentIds,
+} from "@/utils/fileUploadUtils";
 
 export default function CreatePostDrawer({ open, onClose, onSubmit }) {
   const theme = useTheme();
@@ -55,190 +43,38 @@ export default function CreatePostDrawer({ open, onClose, onSubmit }) {
   const companyName = useSelector((state) => state.auth.company.name);
   const projectSteps = useSelector((state) => state.projectStep.items) || [];
 
-  const [form, setForm] = useState({
-    id: "",
-    projectStepId: "",
-    title: "",
-    content: "",
-  });
-  const [files, setFiles] = useState([]); // 첨부 파일 목록
-  const [loadingId, setLoadingId] = useState(false);
+  const {
+    form,
+    setForm,
+    files,
+    setFiles,
+    loadingId,
+    previewModal,
+    handleChange,
+    handleFileSelect,
+    handleFileDelete,
+    handleFileRetry,
+    handleCancel,
+    handlePreviewOpen,
+    handlePreviewClose,
+  } = usePostForm({ dispatch, open });
+
   const [loading, setLoading] = useState(false);
-  const [previewModal, setPreviewModal] = useState({
-    open: false,
-    attachment: null,
-  });
-
-  // 파일 선택 핸들러
-  const handleFileSelect = async (event) => {
-    const selectedFiles = Array.from(event.target.files);
-
-    // 파일 검증
-    const validFiles = [];
-    const invalidFiles = [];
-
-    selectedFiles.forEach((file) => {
-      const errors = validateFile(file);
-      if (errors.length === 0) {
-        validFiles.push(file);
-      } else {
-        invalidFiles.push({ file, errors });
-      }
-    });
-
-    // 검증 실패한 파일들 알림
-    if (invalidFiles.length > 0) {
-      const errorMessages = invalidFiles
-        .map(({ file, errors }) => `${file.name}: ${errors.join(", ")}`)
-        .join("\n");
-
-      alert(`다음 파일들의 업로드가 실패했습니다:\n\n${errorMessages}`);
-    }
-
-    // 중복 파일 체크
-    const newFiles = validFiles.filter(
-      (file) => !files.some((existingFile) => existingFile.name === file.name)
-    );
-
-    if (newFiles.length !== validFiles.length) {
-      const duplicateCount = validFiles.length - newFiles.length;
-      alert(`${duplicateCount}개의 파일이 이미 선택되어 있습니다.`);
-    }
-
-    if (newFiles.length === 0) {
-      return;
-    }
-
-    // 새 파일들을 pending 상태로 추가
-    const newFileItems = newFiles.map((file) => ({
-      id: Date.now() + Math.random(), // 임시 ID
-      file,
-      name: file.name,
-      size: file.size,
-      type: file.type,
-      status: "pending", // pending, uploading, success, error
-      progress: 0,
-      error: null,
-      postAttachmentId: null,
-    }));
-
-    setFiles((prev) => [...prev, ...newFileItems]);
-
-    // 파일들을 순차적으로 업로드 시작
-    for (const fileItem of newFileItems) {
-      await uploadFileWithPresignedUrl({
-        fileItem,
-        postId: form.id,
-        updateFile: (id, updated) => {
-          setFiles((prev) =>
-            prev.map((file) =>
-              file.id === id ? { ...file, ...updated } : file
-            )
-          );
-        },
-      });
-    }
-  };
-
-  // 파일 삭제 핸들러
-  const handleFileDelete = async (fileId) => {
-    const fileToDelete = files.find((file) => file.id === fileId);
-    if (!fileToDelete) return;
-
-    // 확인 대화상자
-    if (!window.confirm(`"${fileToDelete.name}" 파일을 삭제하시겠습니까?`)) {
-      return;
-    }
-
-    try {
-      // 업로드가 완료된 파일인 경우 서버에서도 삭제
-      if (fileToDelete.status === "success" && fileToDelete.postAttachmentId) {
-        await dispatch(
-          deleteAttachment(fileToDelete.postAttachmentId)
-        ).unwrap();
-        console.log(
-          "서버에서 첨부파일 삭제 완료:",
-          fileToDelete.postAttachmentId
-        );
-      }
-
-      // 로컬 상태에서 파일 제거
-      setFiles((prev) => prev.filter((file) => file.id !== fileId));
-
-      // Object URL 정리 (메모리 누수 방지)
-      if (fileToDelete.file) {
-        URL.revokeObjectURL(URL.createObjectURL(fileToDelete.file));
-      }
-    } catch (error) {
-      console.error("파일 삭제 실패:", error);
-      alert("파일 삭제에 실패했습니다. 다시 시도해주세요.");
-    }
-  };
-
-  // 파일 재업로드 핸들러
-  const handleFileRetry = async (fileId) => {
-    const fileItem = files.find((file) => file.id === fileId);
-    if (fileItem) {
-      await uploadFileWithPresignedUrl({
-        fileItem,
-        postId: form.id,
-        updateFile: (id, updated) => {
-          setFiles((prev) =>
-            prev.map((file) =>
-              file.id === id ? { ...file, ...updated } : file
-            )
-          );
-        },
-      });
-    }
-  };
-
-  useEffect(() => {
-    const generateNewPostId = async () => {
-      // 게시글 생성창이 열릴 때마다 새로운 UUID 발급 (독립성 보장)
-      try {
-        setLoadingId(true);
-        const result = await dispatch(createPostId()).unwrap();
-        const newId = typeof result === "string" ? result : result.postId;
-        setForm((prev) => ({ ...prev, id: newId }));
-        console.log("새로운 게시글 UUID 발급:", newId);
-      } catch (err) {
-        console.error("Post ID 생성 실패:", err);
-      } finally {
-        setLoadingId(false);
-      }
-    };
-
-    if (open) {
-      generateNewPostId();
-    }
-  }, [dispatch, open]);
-
-  const handleChange = (key) => (e) =>
-    setForm((f) => ({ ...f, [key]: e.target.value }));
 
   const handleSubmit = async () => {
     const { id, projectStepId, title, content } = form;
     if (!projectStepId || !title.trim() || !content.trim()) return;
 
-    // 파일 업로드가 완료되지 않은 파일이 있는지 확인
-    const hasUnfinishedUploads = files.some(
-      (file) => file.status === "uploading" || file.status === "pending"
-    );
-    if (hasUnfinishedUploads) {
+    if (hasUnfinishedUploads(files)) {
       alert("파일 업로드가 완료될 때까지 기다려주세요.");
       return;
     }
 
-    // 실패한 파일이 있는지 확인
-    const hasFailedUploads = files.some((file) => file.status === "error");
-    if (hasFailedUploads) {
+    if (hasFailedUploads(files)) {
       const confirmSubmit = window.confirm(
         "업로드에 실패한 파일이 있습니다. 그래도 게시글을 등록하시겠습니까?"
       );
-      if (!confirmSubmit) {
-        return;
-      }
+      if (!confirmSubmit) return;
     }
 
     setLoading(true);
@@ -251,8 +87,6 @@ export default function CreatePostDrawer({ open, onClose, onSubmit }) {
         companyName,
         authorName: userName,
       };
-
-      // 1. 게시글 생성
       let createdPostId;
       if (onSubmit) {
         const result = await onSubmit(payload);
@@ -264,16 +98,8 @@ export default function CreatePostDrawer({ open, onClose, onSubmit }) {
         createdPostId = result?.id || id;
       }
 
-      // 2. 업로드 성공한 파일들을 일괄 활성화
-      const successfulFiles = files.filter(
-        (file) => file.status === "success" && file.postAttachmentId
-      );
-      if (successfulFiles.length > 0) {
-        console.log(`${successfulFiles.length}개 파일 일괄 활성화 시작...`);
-        const postAttachmentIds = successfulFiles.map(
-          (file) => file.postAttachmentId
-        );
-
+      const postAttachmentIds = getSuccessfulPostAttachmentIds(files);
+      if (postAttachmentIds.length > 0) {
         try {
           await dispatch(
             bulkActivateAttachments({
@@ -281,66 +107,20 @@ export default function CreatePostDrawer({ open, onClose, onSubmit }) {
               postAttachmentIds,
             })
           ).unwrap();
-          console.log("파일 일괄 활성화 완료");
-        } catch (activateError) {
-          console.error("파일 활성화 실패:", activateError);
-          // 파일 활성화 실패는 게시글 생성 성공에 영향을 주지 않음
+        } catch (e) {
+          console.error("파일 활성화 실패:", e);
         }
       }
 
-      // 3. 폼 초기화 및 창 닫기
       setForm({ id: "", projectStepId: "", title: "", content: "" });
       setFiles([]);
       onClose();
-
-      console.log("게시글 생성 완료:", createdPostId);
     } catch (e) {
       console.error("게시글 생성 실패:", e);
       alert("게시글 등록에 실패했습니다.");
     } finally {
       setLoading(false);
     }
-  };
-
-  const handleCancel = async () => {
-    // 업로드된 파일이 있는 경우 정리
-    const hasUploadedFiles = files.some(
-      (file) => file.status === "success" && file.postAttachmentId
-    );
-
-    if (hasUploadedFiles && form.id) {
-      try {
-        console.log("게시글 취소 - 업로드된 파일 정리 시작:", form.id);
-        await dispatch(cleanupPostAttachments(form.id)).unwrap();
-        console.log("업로드된 파일 정리 완료");
-      } catch (error) {
-        console.error("파일 정리 실패:", error);
-        // 정리 실패해도 창은 닫음
-      }
-    }
-
-    setForm((prev) => ({ ...prev, projectStepId: "", title: "", content: "" }));
-    setFiles([]);
-    onClose();
-  };
-
-  // 미리보기 모달 열기
-  const handlePreviewOpen = (file) => {
-    setPreviewModal({
-      open: true,
-      attachment: {
-        fileName: file.name,
-        fileSize: file.size,
-        fileType: file.type,
-        imageUrl: URL.createObjectURL(file.file),
-      },
-    });
-  };
-
-  // 미리보기 모달 닫기
-  const handlePreviewClose = () => {
-    URL.revokeObjectURL(previewModal.attachment?.imageUrl);
-    setPreviewModal({ open: false, attachment: null });
   };
 
   return (
@@ -366,15 +146,12 @@ export default function CreatePostDrawer({ open, onClose, onSubmit }) {
           }}
         >
           <Stack spacing={4} sx={{ flex: 1 }}>
-            {/* 헤더 */}
             <Box>
               <Box
                 sx={{
                   display: "flex",
                   justifyContent: "space-between",
                   alignItems: "center",
-                  mx: 0,
-                  px: 0,
                 }}
               >
                 <Typography variant="h3" fontWeight={600}>
@@ -551,7 +328,6 @@ export default function CreatePostDrawer({ open, onClose, onSubmit }) {
         </Paper>
       </Drawer>
 
-      {/* 이미지 미리보기 모달 */}
       <FilePreviewModal
         open={previewModal.open}
         attachment={previewModal.attachment}
