+++ src/features/project/post/pages/ProjectPostsPage.jsx
@@ -3,8 +3,8 @@ import { Box } from "@mui/material";
 import { useDispatch, useSelector } from "react-redux";
 import { useParams, useSearchParams, useNavigate } from "react-router-dom";
 import SectionTable from "@/components/common/sectionTable/SectionTable";
-import PostDetailDrawer from "../components/PostDetailDrawer";
-import CreatePostDrawer from "../components/CreatePostDrawer";
+import PostDetailDrawer from "./PostDetailDrawer";
+import CreatePostDrawer from "./CreatePostDrawer";
 import { fetchPosts, fetchPostById, createPost } from "../postSlice";
 import { fetchProjectStages } from "../../slices/projectStepSlice";
 import { POST_APPROVAL_STATUS } from "@/utils/statusMaps";
